fix(sql): userrolemapping only delete admin for default user

+++ transactional/userrolemapping.sql
@@ -1,3 +1,3 @@
-DELETE from public.userrolemapping WHERE userid = 'e4833900-8ab9-11eb-b358-798a371f2ced';
+DELETE from public.userrolemapping WHERE userid = 'e4833900-8ab9-11eb-b358-798a371f2ced' AND role = 'ADMIN';
 INSERT INTO public.userrolemapping (userid,orguserlinkid,"role") VALUES
 	 ('e4833900-8ab9-11eb-b358-798a371f2ced',NULL,'ADMIN');
